+++ my-app/src/components/todoList/Item.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import './css/Item.css'
+import '../css/Item.css'
 
 const Item = ({todo, changeTodoStatus, deleteTodo}) => {
